Datashard read iterator improve performance