@kitaisreal 2023-06-14
Before:
```
./ydb -p test_db benchmark -q "SELECT COUNT(*) FROM hits" --threads 32

Queries executed: 32 QPS 0.02820586771 Average query time: 35.453615s

0.000%  35.651583s
10.000% 35.651583s
20.000% 35.651583s
30.000% 35.651583s
40.000% 35.651583s
50.000% 35.651583s
60.000% 35.651583s
70.000% 35.651583s
80.000% 35.651583s
99.000% 35.651583s
99.900% 35.651583s
99.990% 35.651583s
```
After:
```
./ydb -p test_db benchmark -q "SELECT COUNT(*) FROM hits" --threads 32

Queries executed: 32 QPS 0.03417882026 Average query time: 29.257885s

0.000%  0.000000s
10.000% 0.000000s
20.000% 0.000000s
30.000% 0.000000s
40.000% 0.000000s
50.000% 0.000000s
60.000% 0.000000s
70.000% 0.000000s
80.000% 0.000000s
99.000% 0.000000s
99.900% 0.000000s
99.990% 0.000000s
```